# apps/users-service/api/openapi.yaml
openapi: 3.0.3
info:
  title: Users Service API
  version: 0.2.0
  description: Authentication and internal token validation contract.
servers:
  - url: http://localhost:8081
paths:
  /api/v1/auth/login:
    post:
      summary: Authenticate with login and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials or inactive user
  /api/v1/internal/auth/validate:
    post:
      summary: Validate JWT for internal services
      operationId: validateToken
      security:
        - serviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidateTokenRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateTokenResponse'
        '403':
          description: Missing or invalid service credential
components:
  securitySchemes:
    serviceToken:
      type: apiKey
      in: header
      name: X-Service-Token
  schemas:
    LoginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresInSeconds
        - user
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int32
          example: 3600
        user:
          $ref: '#/components/schemas/UserSummary'
    ValidateTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    ValidateTokenResponse:
      type: object
      required:
        - valid
      properties:
        valid:
          type: boolean
        userId:
          type: integer
          format: int64
        login:
          type: string
        roles:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      required:
        - id
        - login
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string